+++ MFTLib/Kernel32.cs
@@ -6,7 +6,7 @@ namespace MFTLib;
 class Kernel32
 {
     [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-    public static extern SafeFileHandle CreateFile(
+    internal static extern SafeFileHandle CreateFile(
         string lpFileName,
         uint dwDesiredAccess,
         uint dwShareMode,
@@ -16,22 +16,22 @@ class Kernel32
         IntPtr hTemplateFile);
 
     [DllImport("kernel32.dll", SetLastError = true)]
-    public static extern bool CloseHandle(SafeHandle hObject);
+    internal static extern bool CloseHandle(SafeHandle hObject);
 
     [DllImport("kernel32.dll", SetLastError = true)]
-    public static extern bool ReadFile(SafeFileHandle hFile, byte[] lpBuffer, uint nNumberOfBytesToRead,
+    internal static extern bool ReadFile(SafeFileHandle hFile, byte[] lpBuffer, uint nNumberOfBytesToRead,
         out uint lpNumberOfBytesRead, IntPtr lpOverlapped);
 
 
     [DllImport("kernel32.dll", SetLastError = true)]
-    public static extern bool SetFilePointerEx(
+    internal static extern bool SetFilePointerEx(
         SafeFileHandle hFile,
         long liDistanceToMove,
         IntPtr lpNewFilePointer,
         uint dwMoveMethod);
 
     [DllImport("kernel32.dll", SetLastError = true)]
-    public static extern bool DeviceIoControl(SafeFileHandle hDevice, uint dwIoControlCode, IntPtr lpInBuffer,
+    internal static extern bool DeviceIoControl(SafeFileHandle hDevice, uint dwIoControlCode, IntPtr lpInBuffer,
         uint nInBufferSize, IntPtr lpOutBuffer, uint nOutBufferSize, out uint lpBytesReturned,
         IntPtr lpOverlapped);
 }