+++ MFTLib/MatchFlags.cs
@@ -1,0 +1,10 @@
+namespace MFTLib;
+
+[Flags]
+public enum MatchFlags : uint
+{
+    None = 0,
+    ExactMatch = 1,
+    Contains = 2,
+    ResolvePaths = 4,
+}